+++ pyabc/smc.py
@@ -25,6 +25,7 @@ from typing import Union
 from .model import SimpleModel
 from .populationstrategy import ConstantPopulationSize
 from .platform_factory import DefaultSampler
+from .acceptor import accept_use_current_time, SimpleAcceptor
 import copy
 import warnings
 
@@ -118,6 +119,12 @@ class ABCSMC:
         will parallelize across the cores of a single
         machine only.
 
+    acceptor: Acceptor, optional
+        Takes a distance function, summary statistics and an epsilon threshold
+        to decide about acceptance of a particle. Argument accepts any subclass
+        of :class:`pyabc.acceptor.Acceptor` or a function convertible to an
+        acceptor.
+
 
     Attributes
     ----------
@@ -146,7 +153,8 @@ class ABCSMC:
                  model_perturbation_kernel: ModelPerturbationKernel = None,
                  transitions: List[Transition] = None,
                  eps: Epsilon = None,
-                 sampler=None):
+                 sampler=None,
+                 acceptor=None):
 
         if not isinstance(models, list):
             models = [models]
@@ -193,6 +201,10 @@ class ABCSMC:
         if sampler is None:
             sampler = DefaultSampler()
         self.sampler = sampler
+
+        if acceptor is None:
+            acceptor = accept_use_current_time
+        self.acceptor = SimpleAcceptor.assert_acceptor(acceptor)
 
         self.stop_if_only_single_model_alive = False
         self.x_0 = None
@@ -505,6 +517,7 @@ class ABCSMC:
                 self.summary_statistics,
                 self.distance_function,
                 self.eps,
+                self.acceptor,
                 self.x_0)
             # append to all_sum_stats in either case to allow for the situation
             # that in population.all_sum_stats() one is only interested in